requirements.txt
llama-index
openai
pydantic
pyyaml
chainlit 